researcher: widen ResearchAccount::m_total_magnitude to uint64_t
ResearchAccount::m_total_magnitude was uint32_t accumulating
CBlockIndex::Magnitude() (a double). The implicit (double -> uint32_t)
cast in `account.m_total_magnitude += pindex->Magnitude();` is UB once
the running sum exceeds UINT_MAX, which UBSan reported at
tally.cpp:288.

Widen the field to uint64_t (UINT64_MAX comfortably exceeds any
realistic lifetime magnitude sum) and make the double-to-integer cast
explicit at the assignment site. The TODO that previously hung above
this block is resolved.

m_total_magnitude / m_accuracy / AverageLifetimeMagnitude back
reporting-only RPC fields ("Lifetime Magnitude Sum", "Lifetime
Magnitude Average", "Lifetime Payments") -- not accrual, not
consensus. Adjust the RPC pushKV for "Lifetime Magnitude Sum" to
emit int64_t so the field actually carries the wider type instead of
narrowing back to int (which was already lossy for long-lived
researchers).

+++ src/gridcoin/tally.cpp
@@ -281,11 +281,15 @@ public:
 
         account.m_total_research_subsidy += pindex->ResearchSubsidy();
 
-        // TODO: This probably doesn't work correctly given the implicit cast and should be removed. It isn't
-        // used in accrual calculations, only reporting.
+        // m_total_magnitude / m_accuracy back the AverageLifetimeMagnitude
+        // reporting only (not accrual / consensus). m_total_magnitude was
+        // historically uint32_t; the implicit (double -> uint32_t) cast in
+        // the += overflowed once the lifetime sum passed UINT_MAX, which
+        // UBSan reports. Widened to uint64_t in account.h, and the
+        // truncating double-to-integer cast is made explicit here.
         if (pindex->Magnitude() > 0) {
             account.m_accuracy++;
-            account.m_total_magnitude += pindex->Magnitude();
+            account.m_total_magnitude += static_cast<uint64_t>(pindex->Magnitude());
         }
 
         if (account.m_first_block_ptr == nullptr) {
